study_guide_agent/requirements.txt
google-adk
opentelemetry-instrumentation-google-genai
mcp
python-dotenv